refactor(index.js): moved configureSocialLogins to a new file
